requirements.txt
python-telegram-bot==20.7
openai==1.51.0
redis==7.0.1
python-dotenv==1.2.1
psycopg==3.2.12